Remove GrProgramDesc's need for the GrGpu
Although not strictly necessary, being able to compute the program key w/o needing access to the Gpu will give us more flexibility.

Bug: skia:9455

+++ src/gpu/vk/GrVkSampler.h
@@ -25,14 +25,14 @@ public:
     const VkSampler* samplerPtr() const { return &fSampler; }
 
     struct Key {
-        Key(uint16_t samplerKey, const GrVkSamplerYcbcrConversion::Key& ycbcrKey) {
+        Key(uint8_t samplerKey, const GrVkSamplerYcbcrConversion::Key& ycbcrKey) {
             // We must memset here since the GrVkSamplerYcbcrConversion has a 64 bit value which may
             // force alignment padding to occur in the middle of the Key struct.
             memset(this, 0, sizeof(Key));
             fSamplerKey = samplerKey;
             fYcbcrKey = ycbcrKey;
         }
-        uint16_t                        fSamplerKey;
+        uint8_t                         fSamplerKey;
         GrVkSamplerYcbcrConversion::Key fYcbcrKey;
 
         bool operator==(const Key& that) const {
